Sept. 27, 2024

+++ source/SL2P_NetsTools.py
@@ -396,7 +396,7 @@ def estimate_VParams(inParams, DS_Options, inImg, netID_map):
     VP_Options = SL2P_V1.make_VP_options(param_name)  #VP => vegetation parameter
     if VP_Options != None:
       estimate = one_vege_param_map(estimateSL2P_2DNets, VP_Options, stacked_data, sub_netID_map)
-      out_veg_maps[param_name] = estimate
+      out_veg_maps[param_name] = estimate.astype(np.float32)
       
       # generate sl2p output product flag
       QC_output = invalidOutput(estimate, VP_Options)*2
@@ -404,7 +404,7 @@ def estimate_VParams(inParams, DS_Options, inImg, netID_map):
 
       #outDF['error'+v_param] = one_vege_param_map(errorsSL2P_2DNets,   VP_Options, DS_Options, inImg, cliped_netID_map)
 
-  out_veg_maps[eoIM.pix_QA] = xr.DataArray(data=QC_img, dims=['y', 'x'])
+  out_veg_maps[eoIM.pix_QA] = xr.DataArray(data=QC_img, dims=['y', 'x']).astype(np.uint8)
 
   return out_veg_maps.where(date_img > 0)
